backend/requirements.txt
# Django and REST framework
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.2
django-cors-headers>=4.0.0
django-filter>=23.2
bleach>=6.1.0

# Database
psycopg2-binary>=2.9.6
django-db-connection-pool>=1.2.0
sqlparse>=0.4.4

# Celery
celery>=5.3.1
redis>=4.6.0
django-redis>=5.3.0
django-celery-beat>=2.5.0

# File handling and security
python-magic-bin>=0.4.14
yara-python>=4.3.0
pyclamd>=0.4.0
pytz>=2023.3
django-ratelimit>=4.1.0
django-csp>=3.7

# VirusTotal API
vt-py>=0.17.5
requests>=2.31.0

# Storage
django-storages>=1.13.2
minio>=7.1.15

# API documentation
drf-yasg>=1.21.6

# PDF generation
reportlab>=4.0.4
xhtml2pdf>=0.2.11

# WebSockets
channels>=4.0.0
channels-redis>=4.1.0
daphne>=4.0.0
websocket-client>=1.7.0

# Testing
pytest>=7.3.1
pytest-django>=4.5.2

# Utilities
python-dotenv>=1.0.0
Pillow>=10.0.0

# PCAP Analysis
scapy>=2.5.0
numpy>=1.26.0,<2.0.0
pandas>=2.0.0,<3.0.0
scikit-learn>=1.4.0
joblib>=1.3.0
scipy>=1.11.0

# Network Analysis
dnspython>=2.4.0
requests>=2.31.0
geoip2>=4.7.0